Update merge.ps1
Verification of the existence of the ./dist folder to avoid the compilation error during the 1st launch

+++ merge.ps1
@@ -1,2 +1,4 @@
+if (Test-Path -LiteralPath dist/ ) {
 Remove-Item dist/singleFile.cs
+}
 ./builder/CSharpSourcesToSingleFile.exe --verbose --output ./dist/singleFile.cs --directory ./app/files
